requirements.txt
anndata==0.6.20
cycler==0.10.0
decorator==4.4.0
h5py==2.9.0
joblib==0.13.2
kiwisolver==1.1.0
llvmlite==0.28.0
matplotlib==3.0.0
matplotlib-venn==0.11.5
mock==3.0.5
natsort==6.0.0
networkx==2.2
numba==0.43.1
numexpr==2.6.9
numpy==1.15.2
pandas==0.23.4
patsy==0.5.1
pyparsing==2.2.2
python-dateutil==2.8.0
pytz==2019.1
scanpy==1.4.3
scikit-learn==0.20.0
scipy==1.1.0
seaborn==0.9.0
six==1.12.0
statsmodels==0.9.0
tables==3.5.1
tqdm==4.32.1
umap-learn==0.3.9
